pyproject.toml
[tool.poetry]
name = "ultrathink"
version = "1.0.0"
description = "ULTRATHINK Pricing Intelligence System"
authors = ["Your Team <team@company.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.1"
requests = "^2.31.0"
pydantic = "^2.5.0"
praw = "^7.7.1"
snscrape = "^0.7.0"
playwright = "^1.40.0"
google-api-python-client = "^2.108.0"
openai = "^1.6.1"
jinja2 = "^3.1.2"
markdown = "^3.5.1"
premailer = "^3.10.0"
pandas = "^2.1.4"
numpy = "^1.26.2"
click = "^8.1.7"
rich = "^13.7.0"
tabulate = "^0.9.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
ipython = "^8.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
python_files = "test_*.py"

[tool.coverage.run]
source = ["fetchers", "summarizer", "email"]
omit = ["*/tests/*", "*/test_*.py"]